azurefunctions/recipatorfuncapp/requirements.txt
mysql-connector
pandas
scikit-learn
flask
flask-cors